backend/requirements.txt
aiojobs==1.1.0
anyio==3.7.0
async-timeout==4.0.2
boto3==1.26.151
botocore==1.29.151
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.1
fastapi-jsonrpc==2.4.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
passlib==1.7.4
peewee==3.16.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
